requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
atlassian-python-api==3.8.0
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
Deprecated==1.2.12
Flask==1.1.2
Flask-Caching==1.10.1
Flask-RESTful==0.3.8
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
PyGithub==1.54.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
python-dotenv==0.17.0
pytz==2021.1
rauth==0.7.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
sh==1.14.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
